Refactored OpenAIApiClient to read the API key from a file
Created a constant projectName to store the name of the project.
Added QFile and QDir header files to OpenAIApiClient.cpp.
Added a constructor that initializes api_key by reading it from a file. The file path is constructed using QDir::homePath() and projectName.
Modified the constructor to call the new constructor that reads the API key from a file.
Modified mainwindow.cpp to pass an empty QString to the OpenAIApiClient constructor.
Removed the api_key parameter from the MainWindow constructor.